Fix Training Module Selection of a Block (#3664)
This replaces the three separate inputs with just one input for adding/removing training modules from a Timeline Block. They still get partitioned into three different types when in view mode.
